All Animation Errors Fixed - Breathable - Loot
I have fixed all the animation errors and sound errors. I have also updated the breathes water and such components in the MB's json files to match what it says in Bridge. The only error I get now is spawning the buff_mb and that is with a loop on his animations, but the animations work fine. Me and Cursor tried to fix this, but it seems it may just be giving the error as a Bedrock quirk in development, although we have everything technically correct. And we fixed a loot table error when killing the infected mb. This was due to there not being a chance that it could drop nothing (the loot being empty), so we added that. Seeing errors should be a lot easier now that we don't have a bunch when we load the world  :D