Spawn Rules and Script
I added a lot of spawn rules. Each number in the world_age_filter is 1 second in game, so we need to test this, but it seemed to work when I put it down to 120 seconds. Also, We need to tweak or go over all other spawn rules to match vanilla, but I tried my best so far.