Day Tracker and Infected Player Inventory
The Day Tracker should be fine, but we need to double check that the spawning is still working. Also the Infected Player Inventory is not tested. Basically, if you get infected and die, your infected body retains a few items, but most are cooked into coke. If you kill your infected dead body, which is just the Maple Bear that spawns when you die from "snow" and possibly other stuff in the future, then you get a few items and snow back.