Spawn Rules and main.js Update
Tried to fix spawning. If you set the world age filter to 2400, they do not spawn, but I set it to 2, and they did. So I set it to 10 and did not test it yet. Will test it more when I have time. Also, I tried the scripting route, but if I figure this out, it should be easier. But I will still use script.