Spawn Rules and World Age Filter And Animations
I determined by deductive reasoning that the world age filter integers are seconds. So min:1800, means 30 min before they spawn. I set animals to max:3600 or an hour. All this is real time. I also hopefully fixed an animation controller error for the infected mb that came up, but we can revert if it breaks stuff, but we need to test in game to see what it did. I only did a little.