Removed Some Testing Features and Changed Messages
The vanilla items that I used for testing the infection system no longer do testing stuff, they are just normal now. They are just commented out though, so we should be able to add them back anytime if need be.

I also updated the messages more to be more funny. Ya know, when infected, when you die, etc.

And I also changed the spawn rules for the Maple Bears for world age. Tiny bros after 1 day, thralls after 2, and buffs after 5. I think it is in seconds? So 1200 seconds is 20 min. so... hopefully this works now.