Messages and Spawn Rules Tweaks
Updated and added color coded messages for day tracking and joining the world based on infection state of the world.

Tweaked and hopefully balanced spawn rules based on gameplay testing.

My footage was clipped do to me not positioning it right in OBS 'sob' but oh well. I fixed it now.