MB Day Variants + "snow: book updated
Added variants for the mb's, except the buff one, for now... They get stronger. We will add more. @MrPoohter Can you fix the spawn eggs and such? I got the textures but the names are horrendous.

Updated the "snow" book with more unlockable info.

Blufire wants us to revise the infection system. He says just have one infection, not two. Also, make it so you die sooner if infected for incentive to find a cure or face the consequences. Also make it so that if you consume "snow" while infected it resets your timer, but makes the timer smaller. So, yah. If I started with 2hrs, and then was down to 30 min, and I ate a "snow", then it would go back up to 1hr and 30min. But it I was at an 1hr and 45min, then it would bring me down to 1hr and 30 min. Or some thing like that, like it could add time for a bit, and make you think it was helping you, but then the time added got smaller and smaller, and then it started taking away time... and just yah.