INFECTED FAMILY :D
Whenever we create any more mobs that are not vanilla they need to be in the infected family. I did this because the infected mbs were attacking the infected pigs. But now their in the same family so they dont. I also did some stuff to the dusted dirt. It now has shadow.